Let the basic shapes use all available space (512×512px)
This fixes display bugs in several browsers. If the icons fill the available space as much as possible, positioning with transformations is more exact.
For example the Dutch sign for level crossing previously had a red rhomb in the center because the transformation was not carried out exactly by the browser.
The inaccuracies are probably the results of rounding when the browser calculates the transformation. These rounding errors seem to be minimized, if the icon fills the available space.