Move the baseline of the font to the center
Now both the ascent and the descent of the font is 256, which means the baseline is vertically centered in all glyphs.
This helps with positioning the glyphs in situations where the baseline matters (e.g. when drawing glyphs with Graphics2d in Java).

I decided against a baseline at the bottom of the character (ascent 512, descent 0) because not all glyphs are filling the complete available vertical space but they are normally aligned to the center of the glyph.