Add check-task to gulp for additional sanity checks
For simply building traffico as before, just run `npm run build`.

If you want additional sanity checks, run `npm run check` instead. This runs the exact same build as the command above, but after that is finished, sanity checks are performed, which let the build fail if they are not completed successfully.
Currently the only check is, if for any given country there are two signs defined that share the same name.