Add watch task
Night gathers, and now my watch begins. Winter is coming… :wink:
Once upon a time there was a watch-task in this repo, but at some point it disappeared.
This is a new one, not perfect but already functional. Just run `npm run watch` and as soon as you change one of the files in `icons/` or `dev/`, the `check`-task will run.
One drawback of this approach is, that the whole watch-task stops, when the `check`-task fails.