[Security fix] Don't print Access token in Travis CI
Previously commits with git tags were leaking the GitHub personal access token to the public when they were built on Travis CI.
This is now fixed by using a ~/.git-credentials  file to store the access token.

The rest of the script that deploys the website should now also be more readable, as the gh-pages branch is now cloned into a separate directory instead of checked out in the same directory.